Extract original repo from DelegateRepository
To access the original repository in classes extending the
DelegateRepository class, the delegate() method needs to be invoked.

However, this method is package protected in Gerrit core.

As a workaround we added a DelegateRepositoryUnwrapper class in the same namespace
of DelegateRepository to get hold of the original repository.

Adding the DelegateRepositoryUnwrapper jar in the Gerrit lib allows the new class
to be picked up by the class loader.

Bug: Issue 15997

Change-Id: Ia3fbdc7e9629072c0019c44fbdfc201aa5b317b9