Collect metrics asynchronously
Submit metrics collection tasks to the UpdateGitMetricsExecutor.
This will allow the parallelization of metrics collection
by tuning the pool size dedicated to the plugin.

The thread pool is currently hardcoded but it will become
configurable [1].

With this implementation the only parallelization allowed
is at MetricsCollector level.

Currently we have 2 MetricsCollectors, hence `2 * numberOfUpdatedProjects`
tasks will concur on the same thread pool.

We could introduce more concurrency on the File System metrics
collection by processing the `Stream<Path>` in parallel.
However, since these metrics are not real-time by design, it is
probably better to avoid being too aggressive on the file system.

[1]: https://bugs.chromium.org/p/gerrit/issues/detail?id=16248

Change-Id: I781a56a27d2abf3cda7a68967ab805d2a46b5349