Add numberofprojects metric
Implement new metric plugins_git_repo_metrics_numberofprojects which
captures the number of projects at collection moment.
The plugin reload mode is changed to restart to allow for the new metric
to reload, as due to the order of the events on the plugin reload
mechanism, the previous version of the plugin is maintained until the
last stage, which invalidates the re-creation of the metric.

Bug: Issue 325029893
Change-Id: I86e8a1f7f35be650ebd6e24ad5ca92723fb89d3d