Replace project cache with GerritApi for listing projects
When rendering the number of projects metrics, use an
inexpensive project list instead of loading all projects
in cache, leveraging an internal user context.

Also honour the grace period configuration of the plugin
so that the new metric will respect the update frequency
configured without generating extra workload on the system.

If the metrics is requested more often, return a memoized
value from the earlier executions.

Change-Id: Ica5d219be22ce1f0839596ff654f0402c7c0e555