Add the ability to access different SCM repositories
For now, this is limited to GitLab, the intention is to extend this
to other Git SCMs too.

GitLab stores repositories under a different path from Gerrit, some
logic is required to calculate the new path.
When using GitLab backend, the project should be set to the GitLab Id
of that project, rather than the name. The metric will be exposed with
the project Id too. Further details on GitLab project ids at [1].

[1]https://docs.gitlab.com/ee/user/project/working_with_projects.html#access-a-project-by-using-the-project-id

Change-Id: Ic5fa65b49dbab255a2ede9c3e5d4f017c219bf18