Add ability to access GitLab repositories
GitLab stores repositories under a different path from Gerrit, some
logic is required to calculate the new path.
When using GitLab backend, the project should be set to the GitLab Id
of that project, rather than the name. The metric will be exposed with
the project Id too. Further details on GitLab project ids at [1].

[1]https://docs.gitlab.com/ee/user/project/working_with_projects.html#access-a-project-by-using-the-project-id

Change-Id: I68928a261e3a93feb7a489d44758d2e71bc729cc