Verify that ref updates are not lost because of grace period
Having a grace period for metrics updates should not impact the overall
validity of the metrics but only affect how often they are updated.

Add a new integration test that tries to add two branches in rapid
sequence to a repository and assert that the result of number of loose
refs should not change regardless of having a grace period configured or
not.

Change-Id: Iad865492f39dd5ca8746f89f16c0b7ebe21f60b6