Amend repository metrics introducing a project_name field
The repository name was wrongly used in the metric name, which is
against the common paradigm used in Gerrit for tracking repo-specific
metrics, similarly to the definition of JGit block cache in
JGitMetricModule.

Use the typical Gerrit pattern of parametrised metrics by the
project_name field and rely on Gerrit's name sanitisation logic.

Bug: Issue 413269103
Change-Id: I42440218812183b1dd3c08becc1b31f8bb000e04