Fix metric registration check for existing metrics
Checking for the presence of existing metrics cannot be done anymore by
full match but should only consider the prefix, excluding the project
name that is now part of the metric field.

Enable the shouldRegisterMetricsOnlyOnce() test again now that the
metrics detection logic has been fixed and it is not possible anymore
to trigger the registration of the same metric twice.

Change-Id: Ia43962025f43c00c9c3acc06649580619199230b