Add a metrics reporter for testing metrics generation
Parametric metrics are populated in a lazy way depending on the values
of the parameters at runtime. The evaluation happens at the reporting
time, therefore testing without an actual reporter won't trigger the
generation logic.

Add an Slf4JReporter to MetricsInitializerIT so that the test gets
executed on the fully generated metrics at runtime.

Before this change, the metrics were re-registered multiple times for
making the tests pass, which compensated for the lack of metrics
collection (registration triggers a forced collection); however, the
testing scenarios were not necessarily impacted by the extra metrics
creation, but at runtime it was, causing errors whenever a project had
the same metrics updated and collected multiple times.

Change-Id: Ib8f739b53a0fabb4c38f09abf859f356169a8ccf