Add a metrics reporter for testing metrics generation
Parametric metrics are populated in a lazy way depending on the values
of the parameters at runtime. The evaluation happens at the reporting
time, therefore testing without an actual reporter won't trigger the
generation logic.

Add an Slf4JReporter to MetricsInitializerIT so that the test gets
executed on the fully generated metrics at runtime.

Change-Id: Ib8f739b53a0fabb4c38f09abf859f356169a8ccf