Fix FakeMetricMaker metrics creation
Following the introduction of the parametric metric
based on the project name, the FakeMetricMaker was not
amended accordingly causing the creation of incorrect runtime
behaviour.

As soon as the FakeMetricMaker start simulating
the correct logic behind the metric generation, the test
shouldRegisterMetricsOnlyOnce() starts failing which is a good
thing: the code in GitRepoMetricsCache was broken in the detection
of existing metrics therefore this failure is expected.

Change-Id: I689f2bf2c2e9d39fc816ca2838cd4e594782ae3f